Move modifier tests from test_composites to new test_modifiers
